requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# Document processing
python-docx==1.1.0
Pillow==10.2.0

# OCR
easyocr==1.7.1

# Vector database
chromadb==0.4.22

# OpenAI
openai==1.12.0

# Utilities
tqdm==4.66.1
numpy==1.26.3
